#2292: Orgchart: separate color-coded demo to its own controller for clarity
