+++ showcase/src/main/webapp/sections/orgchart/colorCoded.xhtml
@@ -22,7 +22,7 @@
 ${showcase:getFileContent('/sections/orgchart/example-colorCoded.xhtml')}
             </ui:define>
             <ui:define name="contentTab2">
-${showcase:getFileContent('/org/primefaces/extensions/showcase/controller/OrgchartController.java')}
+${showcase:getFileContent('/org/primefaces/extensions/showcase/controller/ColorCodedOrgchartController.java')}
             </ui:define>
         </ui:decorate>
     </ui:define>
