+++ showcase/src/main/webapp/sections/orgchart/example-colorCoded.xhtml
@@ -20,12 +20,12 @@
 
 <!-- EXAMPLE-SOURCE-START -->
         <pe:orgchart id="orgChart"
-                    widgetVar="colorWidget"
-                    value="#{orgchartController.colorCodedChart}"
-                    nodeId="id"
-                    nodeContent="title"
-                    nodeTitle="name"
-                    style="height:400px">
+                     widgetVar="colorWidget"
+                     value="#{colorCodedOrgchartController.colorCodedChart}"
+                     nodeId="id"
+                     nodeContent="title"
+                     nodeTitle="name"
+                     style="height:400px">
         </pe:orgchart>
 <!-- EXAMPLE-SOURCE-END -->
